# argus-backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Argus AI — Event Collector API
  version: 2.1.0
  description: |
    REST API for the Argus AI proctoring platform.

    ## Authentication

    All endpoints (except `POST /api/v1/auth/login`) require a Bearer JWT in the
    `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via the login endpoint and expire after 24 hours.

    ## Rate Limits

    | Scope       | Limit             |
    |-------------|-------------------|
    | Global      | 5 000 req/s       |
    | Per-IP      | 100 req/s         |
    | Login       | 5 req/s (Nginx)   |

    Rate-limit responses return HTTP 429 with a `Retry-After` header.

  contact:
    name: Argus AI Engineering
    url: https://argus.ai
servers:
  - url: https://api.argus.ai
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/export/bulk:
    post:
      tags: [Export]
      summary: Create bulk export job
      description: |
        Initiates an async job to produce a TAR.GZ archive containing evidence
        fragments and a forensic manifest (SHA-256 chain) for the specified sessions.

        Poll `GET /api/v1/export/{exportId}` for status. When `status = "completed"`,
        the `downloadUrl` field contains a presigned URL.
      operationId: createExportJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExportRequest'
            example:
              sessionIds:
                - "sess_abc123"
                - "sess_def456"
              presignTtlSec: 3600
      responses:
        '202':
          description: Export job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/export:
    get:
      tags: [Export]
      summary: List export jobs
      operationId: listExportJobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Export job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExportJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/export/{exportId}:
    parameters:
      - name: exportId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Export]
      summary: Get export job status
      operationId: getExportJob
      responses:
        '200':
          description: Export job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Export]
      summary: Cancel export job
      description: Cancels a pending or processing export job.
      operationId: cancelExportJob
      responses:
        '204':
          description: Cancelled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot cancel a completed or already-cancelled job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Appeals ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Parameters ─────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "not found"
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "invalid request body"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CreateExportRequest:
      type: object
      required: [sessionIds]
      properties:
        sessionIds:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 500
        presignTtlSec:
          type: integer
          default: 3600
          description: TTL for the download URL in seconds
    ExportJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
        requestedBy:
          type: string
        sessionIds:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        archiveUri:
          type: string
          nullable: true
        sha256Archive:
          type: string
          nullable: true
          description: SHA-256 of the complete TAR.GZ archive
        downloadUrl:
          type: string
          nullable: true
          description: Presigned download URL (only present when status = completed)
        errorMessage:
          type: string
          nullable: true
        totalSizeBytes:
          type: integer
          format: int64
          nullable: true
        fragmentCount:
          type: integer
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true

    # ── Appeals ─────────────────────────────────────────────────────────────
...
security:
  - BearerAuth: []